feat(api/votes): implements 'DELETE /api/polls/{ID}/votes` for deleting a vote
This also creates a vote status. The initial vote status is pending.

The possible states are `valid`, `pending` and `invalid`.

The summarizer should look for `pending` or `valid` with deleted at.
`pending` should become `valid` and `valid` should become `invalid`.

Future fraud analysis can look votes in all states.

The goal is to the summarize generator to count the votes for a poll.